Update patch set 1
Patch Set 1:

> Shouldn't it be &&?

No, if the locales do not match the current dict should be skipped. They just didn't take into account that mLocale might be null and calling a method on it will generate a NPE. If it is null the dict obviously should be skipped. I actually hit that yesterday so why not fix it. Of course it would be nicer for createDictionaryInfoFromFileAddress() to return null or even for LocaleUtils.constructLocaleFromString() to throw a NPE in that case. But I don't know what the contract is here and I don't feel like going through the numerous uses to check if that would break things elsewhere.

Patch-set: 1