Update patch set 1
Patch Set 1:

> and why is that better then before? :)

Hello. The Patch [Add "more" keys to ALL the keys] is a very old one. Build around kitkat's keyboard design. That reworked patch here matches current keyboard. That helps to memorize special character positions in [?123] menu.
The only binding we lose is [_] but we win [(],[)] and [!].
Yes, its a little matter of taste but i like this rearrangement more for normal typing.

Patch-set: 1